# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: getContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: listId
          in: query
          schema:
            type: ["integer", "null"]
        - name: validationStatus
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Contacts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contacts/import/preview:
    post:
      operationId: previewContactImport
      tags: [contacts]
      summary: Preview CSV import
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewResponse"
  /contacts/import:
    post:
      operationId: importContacts
      tags: [contacts]
      summary: Import contacts from CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                columnMapping:
                  type: string
                listId:
                  type: string
                skipDuplicates:
                  type: string
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResultResponse"
  /contact-lists:
    get:
      operationId: getContactLists
      tags: [contacts]
      summary: List contact lists
      responses:
        "200":
          description: Contact lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactList"
    post:
      operationId: createContactList
      tags: [contacts]
      summary: Create contact list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactListBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
  /contact-lists/{id}:
    get:
      operationId: getContactList
      tags: [contacts]
      summary: Get contact list by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContactList
      tags: [contacts]
      summary: Delete contact list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /providers/config:
    post:
      operationId: saveProviderConfig
      tags: [providers]
      summary: Save provider configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProviderConfigBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderConfig"
  /providers/test:
    post:
      operationId: testProvider
      tags: [providers]
      summary: Test provider connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestProviderBody"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"
  /webhooks/{provider}:
    post:
      operationId: handleWebhook
      tags: [providers]
      summary: Handle provider webhook
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        normalizedPhone:
          type: string
        email:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        customFields:
          type: object
          additionalProperties: true
        validationStatus:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - normalizedPhone
        - tags
        - customFields
        - validationStatus
        - createdAt
        - updatedAt
    ContactsListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - contacts
        - total
        - page
        - limit
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        customFields:
          type: object
          additionalProperties: true
      required:
        - name
        - phone
    UpdateContactBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        tags:
          type: ["array", "null"]
          items:
            type: string
        customFields:
          type: ["object", "null"]
          additionalProperties: true
        validationStatus:
          type: ["string", "null"]
    ContactList:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        contactCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - contactCount
        - createdAt
        - updatedAt
    CreateContactListBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
    ImportPreviewResponse:
      type: object
      properties:
        headers:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        totalRows:
          type: integer
        filename:
          type: string
      required:
        - headers
        - rows
        - totalRows
        - filename
    ImportResultResponse:
      type: object
      properties:
        importId:
          type: integer
        totalRows:
          type: integer
        importedRows:
          type: integer
        duplicateRows:
          type: integer
        invalidRows:
          type: integer
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - importId
        - totalRows
        - importedRows
        - duplicateRows
        - invalidRows
        - errors
...
    ProviderConfig:
      type: object
      properties:
        id:
          type: integer
        providerName:
          type: string
        baseUrl:
          type: ["string", "null"]
        instanceId:
          type: ["string", "null"]
        isActive:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - providerName
        - isActive
        - createdAt
        - updatedAt
    SaveProviderConfigBody:
      type: object
      properties:
        providerName:
          type: string
        baseUrl:
          type: ["string", "null"]
        instanceId:
          type: ["string", "null"]
        apiToken:
          type: ["string", "null"]
        webhookSecret:
          type: ["string", "null"]
        isActive:
          type: string
      required:
        - providerName
    TestProviderBody:
      type: object
      properties:
        providerName:
          type: string
        baseUrl:
          type: ["string", "null"]
        instanceId:
          type: ["string", "null"]
        apiToken:
          type: ["string", "null"]
      required:
        - providerName
    ProviderTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]
      required:
        - success
        - message