# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactBodyCustomFields } from "./createContactBodyCustomFields";

export interface CreateContactBody {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  tags?: string[];
  customFields?: CreateContactBodyCustomFields;
}