# lib/api-zod/src/generated/types/importPreviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImportPreviewResponseRowsItem } from "./importPreviewResponseRowsItem";

export interface ImportPreviewResponse {
  headers: string[];
  rows: ImportPreviewResponseRowsItem[];
  totalRows: number;
  filename: string;
}