# lib/api-zod/src/generated/types/providerConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderConfig {
  id: number;
  providerName: string;
  /** @nullable */
  baseUrl?: string | null;
  /** @nullable */
  instanceId?: string | null;
  isActive: string;
  createdAt: string;
  updatedAt: string;
}