# lib/api-zod/src/generated/types/providerTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  latencyMs?: number | null;
}